Refactor duplicated citation command formatting in CAYW (biblatex and natbib) formatters
